Split native tool card view family
