+++ Sources/QuillCodeApp/QuillCodeToolCardView.swift
@@ -272,487 +272,3 @@ struct QuillCodeToolCardView: View {
         return "\(card.title), \(card.statusAccessibilityLabel), \(card.densityAccessibilityLabel)\(context)"
     }
 }
-
-private struct QuillCodeToolCardActionRow: View {
-    var actions: [ToolCardActionSurface]
-    var onAction: (ToolCardActionSurface) -> Void
-
-    var body: some View {
-        HStack(spacing: 8) {
-            ForEach(actions) { action in
-                Button {
-                    onAction(action)
-                } label: {
-                    Label {
-                        Text(action.title)
-                    } icon: {
-                        if let systemImage = action.systemImage {
-                            Image(systemName: systemImage)
-                        }
-                    }
-                    .font(.caption.weight(.semibold))
-                    .lineLimit(1)
-                    .padding(.horizontal, 12)
-                    .frame(minHeight: QuillCodeMetrics.minimumHitTarget)
-                    .frame(
-                        minWidth: action.style == .primary ? 118 : 72,
-                        maxWidth: action.style == .primary ? .infinity : 92
-                    )
-                    .foregroundStyle(foregroundColor(for: action.style))
-                    .background(backgroundColor(for: action.style))
-                    .overlay(
-                        Capsule()
-                            .stroke(strokeColor(for: action.style), lineWidth: 1)
-                    )
-                    .clipShape(Capsule())
-                    .contentShape(Capsule())
-                }
-                .buttonStyle(QuillCodePressableButtonStyle())
-                .help(action.title)
-                .accessibilityLabel(action.title)
-                .layoutPriority(action.style == .primary ? 1 : 0)
-            }
-            Spacer(minLength: 0)
-        }
-        .padding(.top, 2)
-    }
-
-    private func foregroundColor(for style: ToolCardActionStyle) -> Color {
-        switch style {
-        case .primary:
-            return Color.white
-        case .secondary:
-            return QuillCodePalette.text
-        case .destructive:
-            return QuillCodePalette.red
-        }
-    }
-
-    private func backgroundColor(for style: ToolCardActionStyle) -> Color {
-        switch style {
-        case .primary:
-            return QuillCodePalette.blue
-        case .secondary:
-            return QuillCodePalette.selection.opacity(0.55)
-        case .destructive:
-            return QuillCodePalette.red.opacity(0.14)
-        }
-    }
-
-    private func strokeColor(for style: ToolCardActionStyle) -> Color {
-        switch style {
-        case .primary:
-            return Color.white.opacity(0.14)
-        case .secondary:
-            return Color.white.opacity(0.10)
-        case .destructive:
-            return QuillCodePalette.red.opacity(0.26)
-        }
-    }
-}
-
-private struct QuillCodeToolStatusBadge: View {
-    var label: String
-    var accessibilityLabel: String
-    var tint: Color
-    var iconName: String
-
-    var body: some View {
-        Label(label, systemImage: iconName)
-            .font(.caption.monospacedDigit().weight(.semibold))
-            .lineLimit(1)
-            .padding(.horizontal, 9)
-            .padding(.vertical, 5)
-            .foregroundStyle(tint)
-            .background(tint.opacity(0.15))
-            .overlay(
-                Capsule()
-                    .stroke(tint.opacity(0.22), lineWidth: 1)
-            )
-            .clipShape(Capsule())
-            .help(label)
-            .accessibilityLabel("Tool status \(accessibilityLabel)")
-    }
-}
-
-struct QuillCodeExecutionContextChip: View {
-    var context: ExecutionContextSurface
-
-    var body: some View {
-        HStack(spacing: 5) {
-            Image(systemName: iconName)
-                .font(.caption2.weight(.bold))
-            Text(title)
-                .lineLimit(1)
-        }
-        .font(.caption2.weight(.bold))
-        .foregroundStyle(tint)
-        .padding(.horizontal, 7)
-        .padding(.vertical, 4)
-        .background(background)
-        .overlay(
-            Capsule()
-                .stroke(tint.opacity(context.kind == .sshRemote ? 0.38 : 0.24), lineWidth: 1)
-        )
-        .clipShape(Capsule())
-        .accessibilityLabel("\(context.label) \(context.detail)")
-    }
-
-    private var title: String {
-        switch context.kind {
-        case .local:
-            return context.label
-        case .sshRemote:
-            return "\(context.label) · \(context.detail)"
-        }
-    }
-
-    private var iconName: String {
-        switch context.kind {
-        case .local:
-            return "desktopcomputer"
-        case .sshRemote:
-            return "point.3.connected.trianglepath.dotted"
-        }
-    }
-
-    private var tint: Color {
-        switch context.kind {
-        case .local:
-            return QuillCodePalette.muted
-        case .sshRemote:
-            return QuillCodePalette.purple
-        }
-    }
-
-    private var background: Color {
-        switch context.kind {
-        case .local:
-            return Color.white.opacity(0.07)
-        case .sshRemote:
-            return QuillCodePalette.purple.opacity(0.16)
-        }
-    }
-}
-
-struct QuillCodeExecutionRail: View {
-    var context: ExecutionContextSurface
-
-    var body: some View {
-        Rectangle()
-            .fill(tint.opacity(context.kind == .sshRemote ? 0.78 : 0.42))
-            .frame(width: 3)
-            .padding(.vertical, 8)
-            .clipShape(RoundedRectangle(cornerRadius: 2, style: .continuous))
-            .padding(.leading, 1)
-            .accessibilityHidden(true)
-    }
-
-    private var tint: Color {
-        switch context.kind {
-        case .local:
-            return QuillCodePalette.muted
-        case .sshRemote:
-            return QuillCodePalette.purple
-        }
-    }
-}
-
-private struct QuillCodeArtifactChip: View {
-    var artifact: ToolArtifactState
-
-    var body: some View {
-        Group {
-            if let url = artifactURL {
-                Link(destination: url) {
-                    label
-                }
-            } else {
-                label
-            }
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel("Artifact \(artifact.label)")
-    }
-
-    private var label: some View {
-        HStack(spacing: 6) {
-            Image(systemName: iconName)
-            VStack(alignment: .leading, spacing: 1) {
-                Text(artifact.label)
-                    .lineLimit(1)
-                Text(artifact.detail)
-                    .font(.caption2)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .lineLimit(1)
-            }
-        }
-        .font(.caption.weight(.semibold))
-        .foregroundStyle(QuillCodePalette.blue)
-        .frame(maxWidth: 260, alignment: .leading)
-        .frame(minHeight: 40)
-        .padding(.horizontal, 10)
-        .padding(.vertical, 6)
-        .background(QuillCodePalette.blue.opacity(0.12))
-        .overlay(
-            Capsule()
-                .stroke(QuillCodePalette.blue.opacity(0.28), lineWidth: 1)
-        )
-        .clipShape(Capsule())
-    }
-
-    private var artifactURL: URL? {
-        artifact.href.flatMap(URL.init(string:))
-    }
-
-    private var iconName: String {
-        if let documentPreview = artifact.documentPreview {
-            return documentPreview.systemImage
-        }
-        switch artifact.kind {
-        case .url:
-            return "link"
-        case .file:
-            return "doc.text"
-        case .path:
-            return "folder"
-        }
-    }
-}
-
-private struct QuillCodeArtifactDocumentPreview: View {
-    var artifact: ToolArtifactState
-
-    var body: some View {
-        Group {
-            if let url = artifactURL {
-                Link(destination: url) {
-                    content
-                }
-            } else {
-                content
-            }
-        }
-        .buttonStyle(.plain)
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel(accessibilityLabel)
-    }
-
-    private var content: some View {
-        HStack(alignment: .center, spacing: 12) {
-            ZStack {
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .fill(QuillCodePalette.blue.opacity(0.14))
-                Image(systemName: preview?.systemImage ?? "doc")
-                    .font(.title3.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.blue)
-                    .accessibilityHidden(true)
-            }
-            .frame(width: 44, height: 52)
-            .overlay(
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .stroke(Color.white.opacity(0.10), lineWidth: 1)
-            )
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(typeLine)
-                    .font(.caption2.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.blue)
-                    .lineLimit(1)
-                Text(artifact.label)
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.text)
-                    .lineLimit(1)
-                Text(preview?.detail ?? artifact.detail)
-                    .font(.caption2)
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .lineLimit(1)
-            }
-            Spacer(minLength: 4)
-            if artifactURL != nil {
-                Image(systemName: "arrow.up.right")
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-                    .accessibilityHidden(true)
-            }
-        }
-        .padding(10)
-        .frame(maxWidth: .infinity, minHeight: 74, alignment: .leading)
-        .quillCodeSurface(
-            fill: Color.white.opacity(0.05),
-            radius: 18,
-            stroke: Color.white.opacity(0.08),
-            shadow: false
-        )
-    }
-
-    private var preview: ToolArtifactDocumentPreview? {
-        artifact.documentPreview
-    }
-
-    private var typeLine: String {
-        guard let preview else { return "Document" }
-        return "\(preview.typeLabel) · \(preview.extensionLabel)"
-    }
-
-    private var artifactURL: URL? {
-        artifact.href.flatMap(URL.init(string:))
-    }
-
-    private var accessibilityLabel: String {
-        "\(typeLine) preview \(artifact.label)"
-    }
-}
-
-private struct QuillCodeArtifactImagePreview: View {
-    var artifact: ToolArtifactState
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            if let url = previewURL {
-                AsyncImage(url: url) { phase in
-                    switch phase {
-                    case .success(let image):
-                        image
-                            .resizable()
-                            .scaledToFit()
-                    case .failure:
-                        fallback
-                    case .empty:
-                        ProgressView()
-                            .frame(maxWidth: .infinity, minHeight: 120)
-                    @unknown default:
-                        fallback
-                    }
-                }
-                .frame(maxWidth: .infinity)
-                .frame(height: 180)
-                .background(Color.black.opacity(0.22))
-                .quillCodeImageOutline(radius: 10)
-            } else {
-                fallback
-            }
-            VStack(alignment: .leading, spacing: 3) {
-                if let preview = artifact.imagePreview {
-                    Text("\(preview.typeLabel) · \(preview.extensionLabel)")
-                        .font(.caption2.weight(.bold))
-                        .foregroundStyle(QuillCodePalette.blue)
-                        .lineLimit(1)
-                    Text(artifact.label)
-                        .font(.caption.weight(.semibold))
-                        .foregroundStyle(QuillCodePalette.text)
-                        .lineLimit(1)
-                    Text(preview.detail)
-                        .font(.caption2.weight(.semibold))
-                        .foregroundStyle(QuillCodePalette.muted)
-                        .lineLimit(1)
-                } else {
-                    Text(artifact.label)
-                        .font(.caption.weight(.semibold))
-                        .foregroundStyle(QuillCodePalette.text)
-                        .lineLimit(1)
-                }
-            }
-        }
-        .padding(8)
-        .quillCodeSurface(
-            fill: Color.white.opacity(0.05),
-            radius: 18,
-            stroke: Color.white.opacity(0.08),
-            shadow: false
-        )
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel(accessibilityLabel)
-    }
-
-    private var previewURL: URL? {
-        artifact.previewURL.flatMap(URL.init(string:))
-    }
-
-    private var accessibilityLabel: String {
-        guard let preview = artifact.imagePreview else {
-            return "Image preview \(artifact.label)"
-        }
-        return "\(preview.typeLabel) \(preview.extensionLabel) preview \(artifact.label)"
-    }
-
-    private var fallback: some View {
-        VStack(spacing: 8) {
-            Image(systemName: "photo")
-                .font(.title3)
-            Text("Preview unavailable")
-                .font(.caption.weight(.semibold))
-        }
-        .foregroundStyle(QuillCodePalette.muted)
-        .frame(maxWidth: .infinity, minHeight: 120)
-        .background(Color.black.opacity(0.22))
-        .quillCodeImageOutline(radius: 10)
-    }
-}
-
-private struct QuillCodeArtifactTextPreview: View {
-    var artifact: ToolArtifactState
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            HStack(spacing: 8) {
-                Image(systemName: "doc.plaintext")
-                    .foregroundStyle(QuillCodePalette.blue)
-                Text(artifact.label)
-                    .font(.caption.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.text)
-                    .lineLimit(1)
-                Spacer(minLength: 8)
-                Text("Preview")
-                    .font(.caption2.weight(.semibold))
-                    .foregroundStyle(QuillCodePalette.muted)
-            }
-            ScrollView(.horizontal, showsIndicators: false) {
-                Text(artifact.textPreview ?? "")
-                    .font(.system(.caption, design: .monospaced))
-                    .lineLimit(14)
-                    .textSelection(.enabled)
-                    .padding(10)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .background(Color.black.opacity(0.30))
-            .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
-        }
-        .padding(10)
-        .quillCodeSurface(
-            fill: Color.white.opacity(0.05),
-            radius: 18,
-            stroke: Color.white.opacity(0.08),
-            shadow: false
-        )
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel("Text preview \(artifact.label)")
-    }
-}
-
-private struct QuillCodeCodeBlock: View {
-    var title: String
-    var text: String
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 5) {
-            Text(title)
-                .font(.caption.weight(.semibold))
-                .foregroundStyle(QuillCodePalette.muted)
-            ScrollView([.horizontal, .vertical]) {
-                Text(text)
-                    .font(.system(.caption, design: .monospaced))
-                    .textSelection(.enabled)
-                    .padding(10)
-                    .frame(maxWidth: .infinity, alignment: .topLeading)
-            }
-            .frame(maxHeight: QuillCodeMetrics.toolCardRawDetailsMaxHeight, alignment: .topLeading)
-            .background(Color.black.opacity(0.35))
-            .overlay(
-                RoundedRectangle(cornerRadius: 9, style: .continuous)
-                    .stroke(Color.white.opacity(0.08), lineWidth: 1)
-            )
-            .clipShape(RoundedRectangle(cornerRadius: 9, style: .continuous))
-        }
-    }
-}
