+++ docs/DECISIONS.md
@@ -169,7 +169,7 @@
 - MCP resource and prompt discovery now feeds read-only agent tools. If a server advertises `resources` or `prompts` in `initialize`, QuillCode performs bounded optional `resources/list` and `prompts/list` probes, stores resource display names plus URIs, surfaces the counts and first names in Extensions, and ignores probe failures so an otherwise healthy MCP server is not hidden. Ready servers expose `host.mcp.resource.read` and `host.mcp.prompt.get` only for advertised resource identifiers and prompt names, keeping resource/prompt consumption read-risk and separate from append-risk generic `host.mcp.call`.
 - Shared visual primitives live in `QuillCodeDesignSystem.swift`, not in individual feature views or the workspace shell. Palette constants, hit-target metrics, `0.96` press feedback, animation helpers, reusable surfaces, and image outlines should be extended there first so extracted SwiftUI controls stay visually consistent without depending on `WorkspaceSwiftUIView.swift`.
 - Transcript message controls live in `QuillCodeTranscriptMessageView.swift`. User and assistant message bubbles, retry/use-as-draft actions, feedback actions, and the shared transcript copy button should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed. Tool-card and artifact-preview rendering remains the next transcript family to extract.
-- Tool-card rendering lives in `QuillCodeToolCardView.swift`. Tool cards, execution chips and rails, artifact chips/previews, and raw detail blocks should evolve there; `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed and how copy actions are wired.
+- Tool-card composition lives in `QuillCodeToolCardView.swift`, while its subfamilies have focused owners: `QuillCodeToolCardControls.swift` owns action rows, status badges, and shared execution-context chips/rails; `QuillCodeToolArtifactViews.swift` owns artifact chips plus text/document/image previews; `QuillCodeToolCardDetailsView.swift` owns raw JSON detail blocks. `WorkspaceSwiftUIView.swift` should only decide where timeline items are placed and how copy actions are wired.
 - Settings and runtime issue rendering live in `QuillCodeSettingsView.swift`. Authentication settings, developer override controls, Computer Use permission onboarding, diagnostics, and `QuillCodeSettingsDraft` should evolve there; `WorkspaceSwiftUIView.swift` should only present the sheet and apply the resulting settings update.
 - Settings surface contracts live in `QuillCodeSettingsSurface.swift`. `WorkspaceSettingsSurface`, `WorkspaceSettingsUpdate`, Computer Use requirement rows, TrustedRouter sign-in copy, Computer Use permission labels, and compatibility decoding stay beside the settings contract; `WorkspaceSurface.swift` should only carry the aggregate `settings` payload and pass current runtime state into it.
 - Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
